twt-follower-api/requirements.txt
twitter-api-client
fastapi[all]
requests
uvicorn